refactor(InputBar): extract draft state management logic
