+++ src/script/components/ConversationListCell/components/CellDescription/CellDescription.tsx
@@ -22,7 +22,7 @@ import {useMemo} from 'react';
 import cx from 'classnames';
 
 import * as Icon from 'Components/Icon';
-import {DraftState, generateConversationInputStorageKey} from 'Components/InputBar/util/DraftStateUtil';
+import {DraftState, generateConversationInputStorageKey} from 'Components/InputBar/common/draftState/draftState';
 import {useLocalStorage} from 'Hooks/useLocalStorage';
 
 import {iconStyle} from './CellDescription.style';
